Get rid of TyphoonDefinitionBase
