+++ Source/Definition/Internal/TyphoonInjectionDefinition.m
@@ -10,7 +10,7 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 #import "TyphoonInjectionDefinition.h"
-#import "TyphoonDefinitionBase+Internal.h"
+#import "TyphoonDefinition+Internal.h"
 
 
 @implementation TyphoonInjectionDefinition
