+++ Source/Definition/TyphoonDefinition.h
@@ -12,7 +12,6 @@
 
 #import <Foundation/Foundation.h>
 
-#import "TyphoonDefinitionBase.h"
 #import "TyphoonMethod.h"
 #import "TyphoonDefinitionNamespace.h"
 
@@ -20,12 +19,55 @@
 @class TyphoonRuntimeArguments;
 @class TyphoonFactoryDefinition;
 
+/**
+ * @ingroup Definition
+ * Describes the lifecycle of a Typhoon component.
+ *
+ * <strong>TyphoonScopeObjectGraph</strong>
+ * (default) This scope is essential (and unique to Typhoon) for mobile and desktop applications. When a component is resolved, any
+ * dependencies with the object-graph will be treated as shared instances during resolution. Once resolution is complete they are not
+ * retained by the TyphoonComponentFactory. This allows instantiating an entire object graph for a use-case (say for a ViewController), and
+ * then discarding it when that use-case has completed, therefore making efficient use of memory.
+ *
+ * <strong>TyphoonScopePrototype</strong>
+ * Indicates that a new instance should always be created by Typhoon, whenever this component is obtained from an assembly or referenced by
+ * another component.
+ *
+ * <strong>TyphoonScopeSingleton</strong>
+ * Indicates that Typhoon should retain the instance that exists for as long as the TyphoonComponentFactory exists.
+ *
+ * <strong>TnglyphoonScopeLazySieton</strong>
+ * This scope behaves the same as TyphoonScopeSingleton, but the object is not created unless or until it is needed.
+ *
+ * <strong>TyphoonScopeWeakSingleton</strong>
+ * Indicates that a shared instance should be created as long as necessary. When your application's classes stop referencing this component,
+ * it will be deallocated until needed again.
+ */
+typedef NS_ENUM(NSInteger, TyphoonScope)
+{
+    TyphoonScopeObjectGraph,
+    TyphoonScopePrototype,
+    TyphoonScopeSingleton,
+    TyphoonScopeLazySingleton,
+    TyphoonScopeWeakSingleton,
+};
+
+// TODO: doc
+typedef NS_OPTIONS(NSInteger, TyphoonAutoInjectVisibility)
+{
+    TyphoonAutoInjectVisibilityNone = 0,
+    TyphoonAutoInjectVisibilityByClass = 1 << 0,
+    TyphoonAutoInjectVisibilityByProtocol = 1 << 1,
+    TyphoonAutoInjectVisibilityDefault = TyphoonAutoInjectVisibilityByClass | TyphoonAutoInjectVisibilityByProtocol,
+};
+
+
 typedef void(^TyphoonDefinitionBlock)(TyphoonDefinition *definition);
 
 /**
 * @ingroup Definition
 */
-@interface TyphoonDefinition : TyphoonDefinitionBase
+@interface TyphoonDefinition : NSObject <NSCopying>
 {
     // TODO: cleanup these
     
@@ -38,21 +80,31 @@ typedef void(^TyphoonDefinitionBlock)(TyphoonDefinition *definition);
     TyphoonDefinition *_parent;
 }
 
-/**
-* A custom callback methods that is invoked before properties and method injection occurs.
-*/
-
-- (void)performBeforeInjections:(SEL)sel;
-- (void)performBeforeInjections:(SEL)sel parameters:(void(^)(TyphoonMethod *params))parametersBlock;
-
+@property (nonatomic, readonly) Class type;
 
 /**
-* A custom callback methods that is invoked after properties and method injection occurs.
-*/
+ * The scope of the component, default being TyphoonScopeObjectGraph.
+ */
+@property (nonatomic) TyphoonScope scope;
 
-- (void)performAfterInjections:(SEL)sel;
-- (void)performAfterInjections:(SEL)sel parameters:(void(^)(TyphoonMethod *param))parameterBlock;
+/**
+ * Specifies visibility for AutoInjection.
+ *
+ * AutoInjection is performed when using method:
+ * - (void)injectProperty:(SEL)withSelector;
+ * or when using:
+ * InjectedClass or InjectedProtocol macro
+ */
+@property (nonatomic) TyphoonAutoInjectVisibility autoInjectionVisibility;
 
+/**
+ * The namespace of the component.
+ */
+@property (nonatomic, readonly) TyphoonDefinitionNamespace *space;
+
+- (void)applyGlobalNamespace;
+
+- (void)applyConcreteNamespace:(NSString *)key;
 
 /**
 * A parent component. When parent is defined the initializer and/or properties from a definition are inherited, unless overridden. Example:
@@ -191,6 +243,18 @@ definition.factory = [self sqliteManager];
 *
 */
 - (void)useInitializer:(SEL)selector;
+
+/**
+ * A custom callback methods that is invoked before properties and method injection occurs.
+ */
+- (void)performBeforeInjections:(SEL)sel;
+- (void)performBeforeInjections:(SEL)sel parameters:(void (^)(TyphoonMethod *params))parametersBlock;
+
+/**
+ * A custom callback methods that is invoked after properties and method injection occurs.
+ */
+- (void)performAfterInjections:(SEL)sel;
+- (void)performAfterInjections:(SEL)sel parameters:(void (^)(TyphoonMethod *params))parameterBlock;
 
 #pragma mark Making injections from definition
