+++ Source/Typhoon.h
@@ -24,7 +24,6 @@ FOUNDATION_EXPORT const unsigned char TyphoonVersionString[];
 #import "TyphoonDefinition.h"
 #import "TyphoonBlockDefinition.h"
 #import "TyphoonFactoryDefinition.h"
-#import "TyphoonDefinitionBase+Infrastructure.h"
 #import "TyphoonDefinition+Infrastructure.h"
 #import "TyphoonMethod.h"
 #import "TyphoonConfigPostProcessor.h"
