+++ Tests/Factory/Internal/TyphoonComponentDefinition+InstanceBuilderTests.m
@@ -13,7 +13,7 @@
 #import "Typhoon.h"
 #import "Knight.h"
 #import "CampaignQuest.h"
-#import "TyphoonDefinitionBase+Internal.h"
+#import "TyphoonDefinition+Internal.h"
 #import "TyphoonDefinition+InstanceBuilder.h"
 #import "PrimitiveMan.h"
 #import "TyphoonInjections.h"
